Update index.html
updated the position of expenses form, this is more effective when lot of exoenses are there and when we scroll that form goes up which does not suits with ui ux. i think this is also the problem when we go to see comment or other videos in yt the main video goes up.